+++ src/code/DCUIRatingScaleHalfStrategy.m
@@ -29,12 +29,6 @@
 }
 
 -(float) calcRatingFromTouchX:(int) touchX{
-	DC_LOG(@"%f", (float)touchX + 1);
-	DC_LOG(@"%f", ((float)touchX + 1)/ (float)imageWidth);
-	DC_LOG(@"%f", ((float)touchX + 1)/ (float)imageWidth * 2);
-	DC_LOG(@"%f", ceilf(((float)touchX + 1) / (float)imageWidth * 2));
-	DC_LOG(@"%f", ceilf(((float)touchX + 1) / (float)imageWidth * 2) / 2);
-
 	return touchX < zeroAreaWidth ? 0 : ceilf(((float)touchX + 1) / (float)imageWidth * 2) /2;
 }