+++ src/tests/DCUIRatingScaleHalfStrategyTests.m
@@ -12,7 +12,7 @@
 #import "DCUIRatingScaleHalfStrategy.h"
 #import "DCCommon.h"
 #import "DCUIRating.h"
-#import "GHUnitTest+DCUIRatingScaleUtils.h"
+#import "GHUnitTest+DCUIRatingFixtures.h"
 
 @interface DCUIRatingScaleHalfStrategyTests : GHTestCase {
 	
@@ -23,17 +23,17 @@
 
 -(void) testDrawImageAtIndexWithRating0 {
 	int imageTypes[5] = {IMAGE_OFF, IMAGE_OFF, IMAGE_OFF, IMAGE_OFF, IMAGE_OFF};
-	[self runDrawAtTestWithRating:0 scaleType:DC_SCALE_0_TO_5_WITH_HALVES imageTypes:imageTypes];
+	[self verifyImagesDrawAtCorrectPostionWithRating:0 scaleType:DC_UI_RATING_SCALE_HALF imageTypes:imageTypes iconCount:5];
 }
 
 -(void) testDrawImageAtIndexWithRating3Half {
 	int imageTypes[5] = {IMAGE_ON, IMAGE_ON, IMAGE_ON, IMAGE_HALF_ON, IMAGE_OFF};
-	[self runDrawAtTestWithRating:3.5 scaleType:DC_SCALE_0_TO_5_WITH_HALVES imageTypes:imageTypes];
+	[self verifyImagesDrawAtCorrectPostionWithRating:3.5 scaleType:DC_UI_RATING_SCALE_HALF imageTypes:imageTypes iconCount:5];
 }
 
 -(void) testDrawImageAtIndexWithRating5 {
 	int imageTypes[5] = {IMAGE_ON, IMAGE_ON, IMAGE_ON, IMAGE_ON, IMAGE_ON};
-	[self runDrawAtTestWithRating:5 scaleType:DC_SCALE_0_TO_5_WITH_HALVES imageTypes:imageTypes];
+	[self verifyImagesDrawAtCorrectPostionWithRating:5 scaleType:DC_UI_RATING_SCALE_HALF imageTypes:imageTypes iconCount:5];
 }
 
 -(void) testCalcNewRatingFromTouchX {